+++ pczt/src/roles/low_level_signer/mod.rs
@@ -33,6 +33,27 @@ impl Signer {
         Ok(Self { pczt })
     }
 
+    /// Exposes the capability to sign the Ironwood spends.
+    #[cfg(all(feature = "orchard", zcash_unstable = "nu7"))]
+    pub fn sign_ironwood_with<E, F>(self, f: F) -> Result<Self, E>
+    where
+        E: From<orchard::pczt::ParseError>,
+        F: FnOnce(&Pczt, &mut orchard::pczt::Bundle, &mut u8) -> Result<(), E>,
+    {
+        let mut pczt = self.pczt;
+
+        let mut tx_modifiable = pczt.global.tx_modifiable;
+
+        let mut bundle = pczt.ironwood.clone().into_parsed()?;
+
+        f(&pczt, &mut bundle, &mut tx_modifiable)?;
+
+        pczt.global.tx_modifiable = tx_modifiable;
+        pczt.ironwood = crate::orchard::Bundle::serialize_from(bundle);
+
+        Ok(Self { pczt })
+    }
+
     /// Exposes the capability to sign the Sapling spends.
     #[cfg(feature = "sapling")]
     pub fn sign_sapling_with<E, F>(self, f: F) -> Result<Self, E>