+++ pczt/src/roles/prover/orchard.rs
@@ -10,6 +10,8 @@ impl super::Prover {
             transparent,
             sapling,
             orchard,
+            #[cfg(zcash_unstable = "nu7")]
+            ironwood,
         } = self.pczt;
 
         let mut bundle = orchard.into_parsed().map_err(OrchardError::Parser)?;
@@ -24,6 +26,35 @@ impl super::Prover {
                 transparent,
                 sapling,
                 orchard: crate::orchard::Bundle::serialize_from(bundle),
+                #[cfg(zcash_unstable = "nu7")]
+                ironwood,
+            },
+        })
+    }
+
+    #[cfg(zcash_unstable = "nu7")]
+    pub fn create_ironwood_proof(self, pk: &ProvingKey) -> Result<Self, OrchardError> {
+        let Pczt {
+            global,
+            transparent,
+            sapling,
+            orchard,
+            ironwood,
+        } = self.pczt;
+
+        let mut bundle = ironwood.into_parsed().map_err(OrchardError::Parser)?;
+
+        bundle
+            .create_proof(pk, OsRng)
+            .map_err(OrchardError::Prover)?;
+
+        Ok(Self {
+            pczt: Pczt {
+                global,
+                transparent,
+                sapling,
+                orchard,
+                ironwood: crate::orchard::Bundle::serialize_from(bundle),
             },
         })
     }